microservices/kafka-utility/app/requirements.txt
# Kafka Utility Service Requirements
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.1
kafka-python>=2.0.2
confluent-kafka>=1.9.2
pydantic>=1.9.0
httpx>=0.22.0
aiokafka>=0.7.2
avro-python3>=1.10.2
